pyproject.toml
[project]
name = "electricity-predictor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "altair>=6.0.0",
    "entsoe-py>=0.7.8",
    "fmiopendata>=0.5.0",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "lightgbm>=4.6.0",
    "mlflow>=3.7.0",
    "optuna>=4.6.0",
    "optuna-integration[xgboost]>=4.6.0",
    "pmdarima>=2.1.1",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "streamlit>=1.52.1",
    "xgboost>=3.1.2",
]
